Refactor to allow readers to be specified at runtime
