+++ app/controllers/hyrax/conference_items_controller.rb
@@ -7,6 +7,8 @@ module Hyrax
     # Adds Hyrax behaviors to the controller.
     include Hyrax::WorksControllerBehavior
     include Hyrax::BreadcrumbsForWorks
+    include HykuAddons::WorksControllerAdditionalMimeTypesBehavior
+
     self.curation_concern_type = ::ConferenceItem
 
     # Use this line if you want to use a custom presenter
