+++ app/services/bolognese/readers/time_based_media_article_reader.rb
@@ -1,0 +1,6 @@
+module Bolognese
+  module Readers
+    class TimeBaseMediaArticleReader < BaseWorkReader
+    end
+  end
+end
